web_widget_digital_signature_sale_order was moved in sale workflow https://github.com/OCA/sale-workflow/pull/252
 Rename module this module in web_widget_digitized_signature
