--- web_widget_digital_signature_sale_order/models/__init__.py
@@ -1,10 +1,0 @@
-# -*- coding: utf-8 -*-
-##############################################################################
-#
-#    Copyright 2015 Lorenzo Battistini - Agile Business Group
-#
-#    License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
-#
-##############################################################################
-
-from . import sale_order
