--- web_widget_digital_signature_sale_order/views/sale_order_view.xml
@@ -1,19 +1,0 @@
-<?xml version="1.0" ?>
-<openerp>
-	<data>
-
-		<record id="view_order_form_signature" model="ir.ui.view">
-            <field name="name">view_order_form_signature</field>
-            <field name="model">sale.order</field>
-            <field name="inherit_id" ref="sale.view_order_form"/>
-            <field name="arch" type="xml">
-            	<xpath expr="//field[@name='note']" position="after">
-                    <div class="oe_clear"/>
-					<label for="signature_image" class="oe_edit_only"/>
-					<h2><field name="signature_image" widget="signature" width="400" height="100"/></h2>
-				</xpath>
-            </field>
-        </record>
-
-	</data>
-</openerp>
