+++ web_widget_digitized_signature/static/description/icon.png
