+++ web_widget_digitized_signature/static/lib/excanvas.js
