+++ web_widget_digitized_signature/static/src/css/digital.css
