+++ web_widget_digitized_signature/static/src/css/jquery-ui.css
