+++ web_widget_digitized_signature/static/src/css/jquery.signature.css
@@ -1,0 +1,5 @@
+/* Styles for signature plugin v1.1.0. */
+.kbw-signature {
+	display: inline-block;
+	border: 1px solid #a0a0a0;
+}
