+++ web_widget_digitized_signature/static/src/img/icon.png
