+++ web_widget_digitized_signature/static/src/js/digital_sign.js
