+++ web_widget_digitized_signature/static/src/xml/digital_sign.xml
